[ocore] Refactored the entire ocore, to ensure String is more efficient and uses Array<char> instead of std::string; though it will REQUIRE small string optimization (oibaker doesn't run); because it allocates onto the heap with small numbers (filling up the chunks).

+++ owc/src/input/binding.cpp
@@ -134,7 +134,7 @@ Binding::Binding(String ostr) {
 			if (end[0] != '#')
 				goto failed;
 
-			u32 num = (u32)end.cutBegin(1).toLong();
+			u32 num = (u32) end.cutBegin(1);
 
 			if (num > 255U)
 				goto failed;
